requirements.txt
asgiref==3.4.1
Babel==2.9.1
boto3==1.18.35
confusable-homoglyphs==3.2.0
Django==3.2.7
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-hijack==3.0.0
django-isbn-field==0.5.2
django-money==2.0.1
django-multiselectfield==0.1.12
django-registration==3.2
django-amazon-ses==4.0.0
python-stdnum==1.16
pytz==2021.1
requests==2.26.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
